requirements.txt
fastapi
uvicorn[standard]
jinja2
azure-identity
azure-mgmt-resource
azure-mgmt-network
azure-mgmt-costmanagement
azure-mgmt-policyinsights
azure-mgmt-security
pyvis
networkx
aiohttp